requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
